+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Apr 10 12:47:19 KST 2024
+gradle.version=7.6.1
