+++ src/main/java/flab/just10minutes/member/controller/AuthController.java
@@ -6,7 +6,6 @@ import flab.just10minutes.member.service.LoginService;
 import flab.just10minutes.member.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,13 +22,13 @@ public class AuthController {
     private final LoginService loginService;
     private final MemberService memberService;
 
-    @PostMapping("/login")
-    public ResponseEntity<HttpStatus> login(@RequestBody @Valid LoginRequest loginRequest) {
-        memberService.existMemberValidate(loginRequest.getId(), loginRequest.getPassword());
-
-        loginService.logIn(memberService.findMemberById(loginRequest.getId()).getUniqueId());
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
+//    @PostMapping("/login")
+//    public ResponseEntity<HttpStatus> login(@RequestBody @Valid LoginRequest loginRequest) {
+//        memberService.existMemberValidate(loginRequest.getId(), loginRequest.getPassword());
+//
+//        loginService.logIn(memberService.findMemberById(loginRequest.getId()).getUniqueId());
+//        return new ResponseEntity<>(HttpStatus.OK);
+//    }
 
     @PostMapping("/logout")
     @MemberLoginCheck
